% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticas.R
\name{calcular_estatisticas_gapminder}
\alias{calcular_estatisticas_gapminder}
\title{Calcular estatisticas de continentes}
\usage{
calcular_estatisticas_gapminder(continentes, anos = 2000)
}
\arguments{
\item{continentes}{Nome do continente (chr)}

\item{anos}{Ano base para calcular as estatisticas (int)}
}
\value{
Uma tabela com as estatisticas de um continente
}
\description{
O objetivo desa funcao e calcular estatisticas dos continentes
}
\examples{
calcular_estatisticas_gapminder("Europa")
calcular_estatisticas_gapminder("Ásia")

}
