% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_ls.R
\name{beta_ls}
\alias{beta_ls}
\title{Evaluate Linear Loss Function}
\usage{
beta_ls(resp, pred, beta, norm = "L2")
}
\arguments{
\item{resp}{A \code{vector} of dimension n.}

\item{pred}{A \code{matrix} containing predictors.}

\item{beta}{A \code{vector} containing coefficients.}

\item{norm}{A \code{character} defining loss to use (default is `L2`).}
}
\value{
A \code{numeric} giving value of loss at \code{beta}
}
\description{
This function delivers the value of the loss function for a given balue of \code{beta}
}
\examples{
set.seed(123)
X <- matrix(rnorm(600), nrow = 200) # true matrix of predictors
epsilon <- rnorm(200, 0, sd = 0.25) # observation error
X <- cbind(rep(1, 200), X) # add a column of ones to generate the response with an intercept value
beta <- c(-1, 4, -5, 2) # true coefficient values we want to estimate (first element is the intercept)
y <- X\%*\%beta + epsilon # generate response y based on this model
beta_ls(resp = y, pred = X, beta = beta, norm = "L2")
}
\author{
Rob Molinari
}
